feat: element 关联 geometry 以及 label
